Added users and allowed them to sign in and sign up and also put restrictions on certain crud

+++ src/components/Operators.vue
@@ -15,6 +15,7 @@
 import OperatorService from '@/services/operatorservice'
 import Vue from 'vue'
 import VueTables from 'vue-tables-2'
+import * as firebase from 'firebase'
 
 Vue.use(VueTables.ClientTable, {compileTemplates: true, filterByColumn: true})
 
@@ -74,37 +75,46 @@ export default {
       this.$router.push('edit')
     },
     deleteOperator: function (id) {
-      this.$swal({
-        title: 'Are you totaly sure?',
-        text: 'You can\'t Undo this action',
-        type: 'warning',
-        showCancelButton: true,
-        confirmButtonText: 'OK Delete it',
-        cancelButtonText: 'Cancel',
-        showCloseButton: true
-        // showLoaderOnConfirm: true
-      }).then((result) => {
-        console.log('SWAL Result : ' + result.value)
-        if (result.value === true) {
-          OperatorService.deleteOperator(id)
-            .then(response => {
-              // JSON responses are automatically parsed.
-              this.message = response.data
-              console.log(this.message)
-              this.loadOperators()
-              // Vue.nextTick(() => this.$refs.vuetable.refresh())
-              this.$swal('Deleted', 'You successfully deleted this Operator ' + JSON.stringify(response.data, null, 5), 'success')
-            })
-            .catch(error => {
-              this.$swal('ERROR', 'Something went wrong trying to Delete ' + error, 'error')
-              this.errors.push(error)
-              console.log(error)
-            })
-        } else {
-          console.log('SWAL Else Result : ' + result.value)
-          this.$swal('Cancelled', 'Please Log in or sign up to delete an operator!', 'info')
-        }
-      })
+      var user = firebase.auth().currentUser
+      if (user) {
+        this.$swal({
+          title: 'Are you totaly sure?',
+          text: 'You can\'t Undo this action',
+          type: 'warning',
+          showCancelButton: true,
+          confirmButtonText: 'OK Delete it',
+          cancelButtonText: 'Cancel',
+          showCloseButton: true
+          // showLoaderOnConfirm: true
+        }).then((result) => {
+          console.log('SWAL Result : ' + result.value)
+          if (result.value === true) {
+            OperatorService.deleteOperator(id)
+              .then(response => {
+                // JSON responses are automatically parsed.
+                this.message = response.data
+                console.log(this.message)
+                this.loadOperators()
+                // Vue.nextTick(() => this.$refs.vuetable.refresh())
+                this.$swal('Deleted', 'You successfully deleted this Operator ' + JSON.stringify(response.data, null, 5), 'success')
+              })
+              .catch(error => {
+                this.$swal('ERROR', 'Something went wrong trying to Delete ' + error, 'error')
+                this.errors.push(error)
+                console.log(error)
+              })
+          } else {
+            console.log('SWAL Else Result : ' + result.value)
+            this.$swal('Cancelled', 'Please Log in or sign up to delete an operator!', 'info')
+          }
+        })
+      } else {
+        this.$swal({
+          title: 'Please Sign up to preform this action?',
+          type: 'warning',
+          showCloseButton: true
+        })
+      }
     }
   }
 }
